+++ WireUI/Sources/WireMoveToFolderUI/Views/FolderPicker/FolderPicker.swift
@@ -63,8 +63,7 @@ public struct FolderPicker: View {
                     NavigationLink {
                         CreateFolder(
                             viewModel: CreateFolderViewModel(
-                                useCase: createFolderUseCase,
-                                isContextMenuAllowed: viewModel.isContextMenuAllowed
+                                useCase: createFolderUseCase
                             ),
                             conversationName: conversationName,
                             onFolderCreated: { [weak viewModel] createdFolder in
@@ -128,8 +127,7 @@ public struct FolderPicker: View {
                     Folder(identifier: UUID(), name: "Personal")
                 ]
             ),
-            updateConversationFolderUseCase: PreviewMoveConversationToFolderUseCase(),
-            isContextMenuAllowed: true
+            updateConversationFolderUseCase: PreviewMoveConversationToFolderUseCase()
         ), createFolderUseCase: PreviewCreateConversationFolderUseCase(), conversationName: "Test"
     )
 }
