+++ WireUI/Sources/WireSettingsUI/Account/BackupImportExport/Import/EnterPasswordView/EnterPasswordView.swift
@@ -24,7 +24,6 @@ struct EnterPasswordView: View {
     @Binding var password: String
     @Binding var passwordIsWrong: Bool
     var focusPasswordFieldOnAppear = true
-    var isContextMenuAllowed: Bool
 
     let continueAction: (_ password: String) -> Void
     let cancelAction: () -> Void
@@ -87,8 +86,7 @@ struct EnterPasswordView: View {
                 password: $password,
                 placeholder: Strings.EnterPassword.TextField.placeholder,
                 placeholderColor: passwordFieldPlaceholderColor,
-                focusOnAppear: focusPasswordFieldOnAppear,
-                isContextMenuAllowed: isContextMenuAllowed
+                focusOnAppear: focusPasswordFieldOnAppear
             )
             .tint(passwordFieldBorderColor)
             .padding(.bottom, 8)
