+++ WireUI/Sources/WireSettingsUI/Account/BackupImportExport/Previews/EnterPasswordPreview.swift
@@ -29,7 +29,6 @@ struct EnterPasswordPreview: View {
                 EnterPasswordView(
                     password: $password,
                     passwordIsWrong: $isPasswordWrong,
-                    isContextMenuAllowed: true,
                     continueAction: { _ in },
                     cancelAction: {}
                 )
