+++ WireUI/Sources/WireSettingsUI/Account/BackupImportExport/Previews/ImportBackupPreview.swift
@@ -24,8 +24,7 @@ func ImportBackupPreview() -> some View {
         ImportBackupView(
             viewModel: ImportBackupViewModel(
                 importBackupUseCase: PreviewImportBackupUseCase(),
-                logger: PreviewLogger(),
-                isContextMenuAllowed: true
+                logger: PreviewLogger()
             )
         )
     }
