+++ WireUI/Tests/WireMoveToFolderUITests/FolderPickerViewModelTests.swift
@@ -155,8 +155,7 @@ final class FolderPickerViewModelTests: XCTestCase {
         sut = FolderPickerViewModel(
             conversation: conversation,
             directory: mockDirectory,
-            updateConversationFolderUseCase: mockUpdateFolderUseCase,
-            isContextMenuAllowed: true
+            updateConversationFolderUseCase: mockUpdateFolderUseCase
         )
     }
 }
