+++ WireUI/Tests/WireSettingsUITests/Account/BackupImportExport/Import/ImportBackupViewModelTest.swift
@@ -53,8 +53,7 @@ final class ImportBackupViewModelTest: XCTestCase {
 
         sut = .init(
             importBackupUseCase: mockImportBackupUseCase,
-            logger: mockLogger,
-            isContextMenuAllowed: true
+            logger: mockLogger
         )
     }
