+++ wire-ios/Wire-iOS Share Extension/Sources/Content/PostContent.swift
@@ -20,6 +20,7 @@ import MobileCoreServices
 import UIKit
 import WireCommonComponents
 import WireShareEngine
+import WireFoundation
 
 /// Content that is shared on a share extension post attempt
 final class PostContent {
