+++ wire-ios/Wire-iOS Share Extension/Sources/View Controllers/ShareExtensionViewController.swift
@@ -29,6 +29,7 @@ import WireDomain
 import WireLinkPreview
 import WireLogging
 import WireShareEngine
+import WireFoundation
 
 typealias Completion = () -> Void
 private let zmLog = ZMSLog(tag: "UI")
