+++ wire-ios/Wire-iOS/Sources/AppDelegate.swift
@@ -25,6 +25,7 @@ import WireCoreCrypto
 import WireCountly
 import WireLogging
 import WireSyncEngine
+import WireFoundation
 
 enum ApplicationLaunchType {
     case unknown
