+++ wire-ios/Wire-iOS/Sources/Helpers/MediaShareRestrictionManager.swift
@@ -19,6 +19,7 @@
 import Foundation
 import WireCommonComponents
 import WireSyncEngine
+import WireFoundation
 
 enum ShareableMediaSource: CaseIterable {
     case camera
