+++ wire-ios/Wire-iOS/Sources/Helpers/syncengine/SessionManager+Convenience.swift
@@ -21,6 +21,7 @@ import avs
 import Foundation
 import WireCommonComponents
 import WireSyncEngine
+import WireFoundation
 
 extension SessionManager {
     static var shared: SessionManager? {
