+++ wire-ios/Wire-iOS/Sources/URLActionRouter.swift
@@ -19,6 +19,7 @@
 import SwiftUI
 import WireCommonComponents
 import WireSyncEngine
+import WireFoundation
 
 extension Notification.Name {
     static let companyLoginDidFinish = Notification.Name("Wire.CompanyLoginDidFinish")
