+++ wire-ios/Wire-iOS/Sources/UserInterface/Components/Views/DragInteractionRestrictionTextView.swift
@@ -18,6 +18,7 @@
 
 import UIKit
 import WireCommonComponents
+import WireFoundation
 
 class DragInteractionRestrictionTextView: TextView {
