+++ wire-ios/Wire-iOS/Sources/UserInterface/Settings/CellDescriptors/SettingsCellDescriptorFactory+Options.swift
@@ -21,6 +21,7 @@ import UIKit
 import WireCommonComponents
 import WireSettingsUI
 import WireSyncEngine
+import WireFoundation
 
 extension SettingsCellDescriptorFactory {
